Refactor small methods into a separate file
